+++ src/WinUI.Markdown/Models/MarkdownViewMode.cs
@@ -1,0 +1,7 @@
+namespace WinUI.Markdown.Controls;
+
+public enum MarkdownViewMode
+{
+    PreviewOnly,
+    DualPane
+}