+++ src/WinUI.Markdown/Models/MonacoEditorTheme.cs
@@ -1,0 +1,8 @@
+namespace WinUI.Markdown.Controls;
+
+public enum MonacoEditorTheme
+{
+    System,
+    Light,
+    Dark
+}